+++ docs/docgen/docgen.config.cjs
@@ -3,6 +3,7 @@ const { defineConfig } = require('vue-docgen-cli')
 const reactConfig = require('./docgen.config.react.cjs')
 const vueConfig = require('./docgen.config.vue.cjs')
 const props = require('./docgen.template.props.cjs')
+const events = require('./docgen.template.events.cjs')
 const component = require('./docgen.template.component.cjs')
 
 module.exports = defineConfig({
@@ -12,6 +13,7 @@ module.exports = defineConfig({
   templates: {
     component,
     props,
+    events,
   },
   ignore: ['**/*.rootstory.tsx', '**/*.cy.tsx'],
   pages: [reactConfig, vueConfig],
